begin reorganizing the project using api folder and flask blueprints
