--- app/Dockerfile
@@ -1,5 +1,0 @@
-FROM python:2-onbuild
-
-EXPOSE 8080
-
-CMD [ "python", "./app.py" ]
