+++ build.sh
@@ -1,2 +1,2 @@
 docker rmi pollease:latest
-docker build --force-rm -t pollease:latest ./app
+docker build --force-rm -t pollease:latest ./pollease
